+++ python/pyspark/sql/tests/coercion/golden_pandas_udf_return_type_coercion_base.csv
@@ -1,0 +1,16 @@
+SQL Type \  Value@Type	[None, None]@list	[True, False]@list	['a', 'b']@list	['12', '34']@list	[Decimal('1'), Decimal('2')]@list	[{'a': 1}, {'b': 2}]@list	[1 2]@ndarray[int8]	[1 2]@ndarray[int16]	[1 2]@ndarray[int32]	[1 2]@ndarray[int64]	[1 2]@ndarray[uint8]	[1 2]@ndarray[uint16]	[1 2]@ndarray[uint32]	[1 2]@ndarray[uint64]	[1. 2.]@ndarray[float16]	[1. 2.]@ndarray[float32]	[1. 2.]@ndarray[float64]	[1.+0.j 2.+0.j]@ndarray[complex64]	[1.+0.j 2.+0.j]@ndarray[complex128]	[array([1, 2, 3], dtype=int32), @list	['1970-01-01T00:00:00.000000000'@ndarray[datetime64[ns]]	['1970-01-01T05:00:00.000000000'@ndarray[datetime64[ns]]	[Timedelta('1 days 00:00:00'), T@list	['A', 'B'] Categories (2, object@Categorical	"{""_1"":{""0"":1,""1"":2}}@Dataframe[_1 int64]"
+boolean	[None, None]	[True, False]	X	X	X	X	[True, True]	[True, True]	[True, True]	[True, True]	[True, True]	[True, True]	[True, True]	[True, True]	X	[True, True]	[True, True]	X	X	X	X	X	X	X	X
+tinyint	[None, None]	[1, 0]	X	[12, 34]	[1, 2]	X	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	X	X	X	X	X	X	X	X
+smallint	[None, None]	[1, 0]	X	[12, 34]	[1, 2]	X	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	X	X	X	X	X	X	X	X
+int	[None, None]	[1, 0]	X	[12, 34]	[1, 2]	X	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	X	X	X	X	X	X	X	X
+bigint	[None, None]	[1, 0]	X	[12, 34]	[1, 2]	X	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	[1, 2]	X	X	X	[0, 86400000000000]	[18000000000000, 104400000000000]	[86400000000000, 172800000000000]	X	X
+string	[None, None]	X	['a', 'b']	['12', '34']	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	['A', 'B']	X
+date	[None, None]	X	X	X	[datetime.date(1970, 1, 2), datetime.dat	X	X	X	[datetime.date(1970, 1, 2), datetime.dat	X	X	X	X	X	X	X	X	X	X	X	[datetime.date(1970, 1, 1), datetime.dat	[datetime.date(1970, 1, 1), datetime.dat	X	X	X
+timestamp	[None, None]	X	X	X	[datetime.datetime(1969, 12, 31, 16, 0, 	X	X	X	X	[datetime.datetime(1969, 12, 31, 16, 0, 	X	X	X	X	X	X	X	X	X	X	[datetime.datetime(1970, 1, 1, 0, 0), da	[datetime.datetime(1970, 1, 1, 5, 0), da	X	X	X
+float	[None, None]	[1.0, 0.0]	X	[12.0, 34.0]	[1.0, 2.0]	X	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	X	X	X	X	X	X	X	X
+double	[None, None]	[1.0, 0.0]	X	[12.0, 34.0]	[1.0, 2.0]	X	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	[1.0, 2.0]	X	X	X	X	X	X	X	X
+array<int>	[None, None]	X	X	[[1, 2], [3, 4]]	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	[[1, 2, 3], [1, 2, 3]]	X	X	X	X	X
+binary	[None, None]	[b'\x01', b'']	[b'a', b'b']	[b'12', b'34']	X	X	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'\x01', b'\x02']	[b'', b'']	[b'', b'']	[b'', b'']	[b'', b'']	[b'', b'']	X	[b'', b'']	[b'', b'']	[b'', b'']	[b'A', b'B']	X
+decimal(10,0)	[None, None]	X	X	X	[Decimal('1'), Decimal('2')]	X	[Decimal('1'), Decimal('2')]	[Decimal('1'), Decimal('2')]	[Decimal('1'), Decimal('2')]	X	[Decimal('1'), Decimal('2')]	[Decimal('1'), Decimal('2')]	[Decimal('1'), Decimal('2')]	X	X	[Decimal('1'), Decimal('2')]	[Decimal('1'), Decimal('2')]	X	X	X	X	X	X	X	X
+map<string,int>	[None, None]	X	X	X	X	[{'a': 1}, {'b': 2}]	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X
+struct<_1:int>	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	X	[Row(_1=1), Row(_1=2)]
